# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Ecommerce Backend API
  version: '1.0'
  description: REST API documentation for the ecommerce backend.
servers:
  - url: http://localhost:4000/api/v1
paths:
...
  /products:
    get:
      tags:
        - Products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List products
    post:
      tags:
        - Products
      requestBody:
        required: true
      responses:
        '201':
          description: Create product
...
  /orders:
    post:
      tags:
        - Orders
      requestBody:
        required: true
      responses:
        '201':
          description: Order created
...